seperated output stuff into seperate file
